+++ pkg/driver/node/mount.go
@@ -18,8 +18,6 @@ limitations under the License.
 package node
 
 import (
-	"bufio"
-	"bytes"
 	"context"
 	"fmt"
 	"os"
@@ -28,6 +26,7 @@ import (
 	"time"
 
 	"github.com/awslabs/aws-s3-csi-driver/pkg/driver/node/awsprofile"
+	"github.com/awslabs/aws-s3-csi-driver/pkg/driver/node/mounter"
 	"github.com/awslabs/aws-s3-csi-driver/pkg/system"
 	"github.com/google/uuid"
 	"k8s.io/klog/v2"
@@ -58,16 +57,6 @@ const (
 
 // https://github.com/awslabs/mountpoint-s3/blob/9ed8b6243f4511e2013b2f4303a9197c3ddd4071/mountpoint-s3/src/cli.rs#L421
 const mountpointDeviceName = "mountpoint-s3"
-
-const (
-	// Due to some reason that we haven't been able to identify, reading `/host/proc/mounts`
-	// fails on newly spawned Karpenter/GPU nodes with "invalid argument".
-	// It's reported that reading `/host/proc/mounts` works after some retries,
-	// and we decided to add retry mechanism until we find and fix the root cause of this problem.
-	// See https://github.com/awslabs/mountpoint-s3-csi-driver/issues/174.
-	procMountsReadMaxRetry     = 3
-	procMountsReadRetryBackoff = 100 * time.Millisecond
-)
 
 type MountCredentials struct {
 	// Identifies how these credentials are obtained.
@@ -158,18 +147,10 @@ type ServiceRunner interface {
 	RunOneshot(ctx context.Context, config *system.ExecConfig) (string, error)
 }
 
-type MountLister interface {
-	ListMounts() ([]mount.MountPoint, error)
-}
-
-type ProcMountLister struct {
-	ProcMountPath string
-}
-
 type S3Mounter struct {
 	Ctx               context.Context
 	Runner            ServiceRunner
-	MountLister       MountLister
+	MountLister       mounter.MountLister
 	MpVersion         string
 	MountS3Path       string
 	kubernetesVersion string
@@ -192,37 +173,11 @@ func NewS3Mounter(mpVersion string, kubernetesVersion string) (*S3Mounter, error
 	return &S3Mounter{
 		Ctx:               ctx,
 		Runner:            runner,
-		MountLister:       &ProcMountLister{ProcMountPath: procMounts},
+		MountLister:       &mounter.ProcMountLister{ProcMountPath: procMounts},
 		MpVersion:         mpVersion,
 		MountS3Path:       MountS3Path(),
 		kubernetesVersion: kubernetesVersion,
 	}, nil
-}
-
-func (pml *ProcMountLister) ListMounts() ([]mount.MountPoint, error) {
-	var (
-		mounts []byte
-		err    error
-	)
-
-	for i := 1; i <= procMountsReadMaxRetry; i += 1 {
-		mounts, err = os.ReadFile(pml.ProcMountPath)
-		if err == nil {
-			if i > 1 {
-				klog.V(4).Infof("Successfully read %s after %d retries", pml.ProcMountPath, i)
-			}
-			break
-		}
-
-		klog.Errorf("Failed to read %s on try %d: %v", pml.ProcMountPath, i, err)
-		time.Sleep(procMountsReadRetryBackoff)
-	}
-
-	if err != nil {
-		return nil, fmt.Errorf("Failed to read %s after %d tries: %w", pml.ProcMountPath, procMountsReadMaxRetry, err)
-	}
-
-	return parseProcMounts(mounts)
 }
 
 // IsMountPoint returns whether given `target` is a `mount-s3` mount.
@@ -403,36 +358,6 @@ func (m *S3Mounter) Unmount(target string) error {
 		klog.V(5).Infof("umount output: %s", output)
 	}
 	return nil
-}
-
-func parseProcMounts(data []byte) ([]mount.MountPoint, error) {
-	var mounts []mount.MountPoint
-
-	scanner := bufio.NewScanner(bytes.NewReader(data))
-	for scanner.Scan() {
-		line := scanner.Text()
-		fields := strings.Fields(line)
-		if len(fields) < 6 {
-			return nil, fmt.Errorf("Invalid line in mounts file: %s", line)
-		}
-
-		mountPoint := mount.MountPoint{
-			Device: fields[0],
-			Path:   fields[1],
-			Type:   fields[2],
-			Opts:   strings.Split(fields[3], ","),
-		}
-
-		// Fields 4 and 5 are Freq and Pass respectively. Ignoring
-
-		mounts = append(mounts, mountPoint)
-	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("Error reading mounts data: %w", err)
-	}
-
-	return mounts, nil
 }
 
 const (
